+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQEYGM4ZAAP32ZTHDM2CWDVE"
+        "01KQFFVBT3F9DT1J8JD33X3KZC"
     )
     r"""The unique identifier of the data source"""