+++ packages/orq-rc/RELEASES.md
@@ -7449,3 +7449,13 @@ Based on:
 - [python v4.9.0-rc.12] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.12 - packages/orq-rc
+
+## 2026-04-30 15:27:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.13] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.13] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.13 - packages/orq-rc